wai-aria/role/region-roles.html

<!doctype html>
<html>
<head>
  <title>Region Role Verification Tests</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/resources/testdriver.js"></script>
  <script src="/resources/testdriver-vendor.js"></script>
  <script src="/resources/testdriver-actions.js"></script>
</head>
<body>

<nav role="region" id="nav_no_label">x</form>
<nav role="region" id="nav_with_label" aria-label="x">x</nav>

<script>

promise_test(async t => {
  const role = await test_driver.get_computed_role(document.getElementById('nav_no_label'));
  assert_equals(role, "navigation", "<nav> region w/o label");
}, "<nav> region w/o label");

promise_test(async t => {
  const role = await test_driver.get_computed_role(document.getElementById('nav_with_label'));
  assert_equals(role, "region", "<nav> region w/ label");
}, "<nav> region w/ label");

</script>
</body>
</html>
